fix(config): restore the two-minute bash default and refresh three stale expectations
tools.bash_timeout_ms still defaulted to the 30 minutes an intermediate commit
introduced, while DEFAULT_BASH_TIMEOUT had already been restored to two minutes,
so the runtime's own pin test had been failing on main.

Also updates two test expectations that outlived deliberate behaviour changes:
the ask_user envelope gained an answered field, and child tool results now render
beside their own call instead of after the batch.

+++ crates/agens-config/src/lib.rs
@@ -521,7 +521,7 @@ pub const SETTINGS: &[SettingSpec] = &[
             minimum: 1_000,
             maximum: 86_400_000,
         },
-        default: SettingValue::Integer(1_800_000),
+        default: SettingValue::Integer(120_000),
         doc: "Timeout for a bash tool invocation.",
     },
     SettingSpec {
